fixed the price stuff

+++ Assets/Scripts/Bank.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class Bank : MonoBehaviour
 {
     public double gold = 0;
-    public int swordValue = 50;
-    public int totalClicks = 0;
+    public double swordValue = 50;
+    public double totalClicks = 0;
 
     // Start is called before the first frame update
     void Start()
